+++ sqlglot/optimizer/qualify.py
@@ -8,12 +8,10 @@ from sqlglot.optimizer.isolate_table_selects import isolate_table_selects
 from sqlglot.optimizer.normalize_identifiers import normalize_identifiers
 from sqlglot.optimizer.qualify_columns import (
     qualify_columns as qualify_columns_func,
+    quote_identifiers as quote_identifiers_func,
     validate_qualify_columns as validate_qualify_columns_func,
 )
 from sqlglot.optimizer.qualify_tables import qualify_tables
-from sqlglot.optimizer.quote_identifiers import (
-    quote_identifiers as quote_identifiers_func,
-)
 from sqlglot.schema import Schema, ensure_schema
