+++ sqlglot/optimizer/qualify_columns.py
@@ -4,8 +4,10 @@ import itertools
 import typing as t
 
 from sqlglot import alias, exp
+from sqlglot._typing import E
+from sqlglot.dialects.dialect import DialectType
 from sqlglot.errors import OptimizeError
-from sqlglot.helper import seq_get
+from sqlglot.helper import case_sensitive, seq_get
 from sqlglot.optimizer.scope import Scope, traverse_scope, walk_in_scope
 from sqlglot.schema import Schema, ensure_schema
 
@@ -411,6 +413,25 @@ def _qualify_outputs(scope):
         new_selections.append(selection)
 
     scope.expression.set("expressions", new_selections)
+
+
+def quote_identifiers(
+    expression: E, dialect: DialectType = None, identify: bool = True, copy: bool = True
+) -> E:
+    """Makes sure all identifiers that need to be quoted are quoted."""
+
+    def _quote(expression: E) -> E:
+        if isinstance(expression, exp.Identifier):
+            name = expression.this
+            expression.set(
+                "quoted",
+                identify
+                or case_sensitive(name, dialect=dialect)
+                or not exp.SAFE_IDENTIFIER_RE.match(name),
+            )
+        return expression
+
+    return expression.transform(_quote, copy=copy)
 
 
 class Resolver:
